requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
supabase==2.23.2
python-dotenv==1.0.0
google-genai==1.46.0

# Development helpers
black==23.9.1
flake8==6.1.0
